// HrPro-Recruit-main/src/types/jobs.types.ts
export type AppliedJob = {
  result: {
    status: Status;
    appliedAt: Date;
    answers: JsonValue;
    job: {
      title: string;
      description: string;
      type: EmployementType;
      category: string;
      workMode: WorkMode;
      currency: Currency;
      skills: string[];
      expired: boolean;
      hasExpiryDate: boolean;
      expiryDate?: Date | null;
      hasSalaryRange: boolean;
      minSalary?: number | null;
      maxSalary?: number | null;
      hasExperiencerange: boolean;
      minExperience?: number | null;
      maxExperience?: number | null;
      isVerifiedJob: boolean;
      deleted: boolean;
      deletedAt?: Date | null;
      postedAt: Date;
      updatedAt: Date;
      responsibilities: string[];
      company: {
        companyName: string;
        companyLogo?: string | null;
        companyEmail: string;
        companyBio: string;
        foundedYear: string;
        numberOfEmployees: string;
        CEOName: string;
        companyType: string;
        city: string;
        country: string;
        website?: string | null;
        linkedinLink?: string | null;
        twitterLink?: string| null;
      }
    };
    zoomMeet:{
      topic:string;
      startTime:Date,
      duration:number,
      joinUrl:string,
      password?:string | null,
      status:string,
      createdAt:Date,
      hostJoinUrl:string,
    }[] | null
}[];
}